Dont rewrite sum of astype to inverse generally
